Add Win10 ad Win11 styling switch
Also, added multiply and divide functionality (should have been a
separate commit...).

When first working on this the I didn't realize there was a significant
different between the Windows 10 and Windows 11 calculator style. For
fun added a switch at the top of the file to be able to toggle between
the two.